# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 学习状态智能助手 API（板块一 + 板块二 MVP）
  version: 1.8.0
  description: |
    中学学情导学与状态评估 API。

    来源：`docs/api-design-unified.md`（唯一生效版本），对应 `PRD-学习状态智能助手.md`。
    本文件严格按该设计文档生成，未引入文档之外的字段。

    主流程：创建目标 ① → 生成计划 ② → 提交学习记录 ③ → 计算状态 ④ → 返回建议 ⑤（外加周期复盘与反馈回路）。

    约定要点：
    - 所有资源以当前 token 用户为作用域，不设 userId 路径参数，不存在跨用户读取接口。
    - 写接口可选携带 `Idempotency-Key` 头，24 小时内重复键直接返回首次结果。
    - 建议与复盘为异步生成：POST 返回 202 + `generation.status = pending`，前端轮询 GET 获取结果。
    - 监护人授权失效时所有写接口返回 403 `GUARDIAN_AUTHORIZATION_EXPIRED`，读接口不受影响。

    阶段口径：pilot。本文档为前后端与 QA 的唯一契约来源。项目当前处于 pilot 阶段（小范围试用验证，非 beta/正式发布），由学生团队维持。板块三契约已于 M1 转正（口径见 docs/module3-m0-decision-proposal.md v1.7 §4.11）。

    **v1.6.0（重构 M0 · 契约冻结）**——本次为重构第一阶段的契约增量，未改动任何既有字段名：
    - **稳定用户 ID 落地（D59）**：`User.userId` 语义明确为稳定内部主键（`u_` 前缀），与登录凭证
      解耦；`AuthMeResponse.user` 增补可选 `userId`（`email` 保留，前端展示仍用它）。
    - **新增 M0 实体 schema**：Exam / Collection / CollectionItem / UsageLedgerEntry / InviteCode /
      ViolationLog / ErrorReport / Medal / UserProfileEntry / TopicSummary / Explanation /
      ChatSession / ChatRawMessage。⚠️ 这些 schema **尚未挂到任何 path**，接口由各板块按里程碑补齐；
      字段名以本节为准，各板块不得另起。
    - **题本升格（D48）**：新增 `ErrorCause` / `ErrorIntent` 两个正交维度，`ErrorRecord` 系列扩展
      `errorCause` / `intent` / `sourceExamId` 三个可选字段（`errorType` 保留为自由文本，向后兼容）。
    - **目标（D6 / D49）**：`Goal` / `GoalSummary` / `GoalCreate` / `GoalUpdate` 增补
      `parentGoalId`（父子树）、`examId` + `targetScore`（引用考试）。`status` 的 enum 未改动。
...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /error-book:
    get:
      tags: [错题本]
      operationId: listErrorBook
      summary: 错题列表
      parameters:
        - name: subject
          in: query
          required: false
          description: 按学科过滤
          schema:
            $ref: '#/components/schemas/Subject'
        - name: status
          in: query
          required: false
          description: 按状态过滤（open/resolved）
          schema:
            type: string
            enum: [open, resolved]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 错题列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBookList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [错题本]
      operationId: createErrorRecord
      summary: 错题录入（原文不出域，异步 embedding + 候选知识点匹配）
      description: |
        原文（rawText/studentAnswer/correctAnswer/errorNote）只在本地落库与本地向量化，
        永不进入出域 payload（EgressGuard knowledge_raw 拒绝序列化，CI 断言）。
        返回 201 + 记录 ID；候选知识点由 GET /knowledge/points/match 另行获取，用户确认后 PATCH 关联。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorRecordCreate'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRecord'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/GuardianAuthorizationExpired'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /error-book/{errorId}:
    get:
      tags: [错题本]
      operationId: getErrorRecord
      summary: 错题详情
      parameters:
        - name: errorId
          in: path
          required: true
          description: 错题 ID（err_ 前缀）
          schema:
            type: string
      responses:
        '200':
          description: 错题详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRecord'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [错题本]
      operationId: updateErrorRecord
      summary: 改错因/状态/关联知识点
      parameters:
        - name: errorId
          in: path
          required: true
          description: 错题 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorRecordUpdate'
      responses:
        '200':
          description: 更新后的错题
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRecord'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [错题本]
      operationId: deleteErrorRecord
      summary: 软删错题（deleted_at，对齐 goals/archive 风格）
      parameters:
        - name: errorId
          in: path
          required: true
          description: 错题 ID
          schema:
            type: string
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRecordDeleted'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /error-book/{errorId}/review:
    post:
      tags: [错题本]
      operationId: reviewErrorRecord
      summary: 提交复习结果（艾宾浩斯间隔复习）
      description: recallCorrect 驱动下一次复习间隔（1/2/4/7/15 天），写入 kb_review_logs 并触发 mastery 重算。
      parameters:
        - name: errorId
          in: path
          required: true
          description: 错题 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSubmit'
      responses:
        '200':
          description: 复习已记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResult'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== 板块二：知识点掌握（PRD 12.3.4） ====================
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: |
        HttpOnly Session Cookie（名为 sid），登录成功后由服务端 Set-Cookie 下发，
        有效期 7 天（见 mock-server SESSION_TTL_MS）。浏览器自动携带，前端无需手动设置。
        鉴权失效时业务接口返回 401 UNAUTHENTICATED。
        （历史说明：早期契约曾假设上游 Bearer token 且不含登录/注册接口，
        实际项目改为自带 /auth/* 会话体系，本契约已据此收录，见「鉴权会话」标签。）
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: 页码，默认 1
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: 每页条数，默认 20，上限 50
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 可选幂等键，24 小时内重复键直接返回首次结果（用于弱网下重复提交学习记录）
      schema:
        type: string
...
  responses:
    ValidationFailed:
      description: 参数缺失或取值越界
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_FAILED
              message: 自评专注度必须为 1-5 的整数
              field: selfReport.focus
    Unauthenticated:
      description: token 缺失或过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: 登录状态已失效，请重新登录
    GuardianAuthorizationExpired:
      description: 监护人授权失效，账号只读（PRD 8.1），仅 GET 可用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: GUARDIAN_AUTHORIZATION_EXPIRED
              message: 监护人授权已过期，请重新确认后继续使用
...
    NotFound:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在
...
    InternalServerError:
      description: 服务端错误。对齐 api-design-unified.md 0.2 节错误码表中的 500 INTERNAL_ERROR。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: 服务开小差了，请稍后再试
...
  schemas:
...
    Subject:
      type: string
      description: 学科代码（拼音首字母大写，与知识点库标准对齐；other 保留兜底）。YW 语文 / SX 数学 / YY 英语 / WL 物理 / HX 化学 / SW 生物 / ZZ 思想政治 / LS 历史 / DL 地理
      enum: [YW, SX, YY, WL, HX, SW, ZZ, LS, DL, other]
...
    Error:
      type: object
      description: 统一错误格式
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 错误码
            message:
              type: string
              description: 可展示的错误说明
            field:
              type: string
              description: 校验失败的字段路径，仅参数校验错误时出现
    Pagination:
      type: object
      description: 列表响应的分页信息
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页条数
        total:
          type: integer
          description: 总条数
...
    ErrorCause:
      type: string
      description: >-
        结构化错因（错题专有，D48）。与 errorType（自由文本，历史兼容）并存：
        errorCause 用于统计与 mastery 归因，errorType 保留自述原文。
        取值对应目标态 §4.4.1 的错因分类；careless 与既有自由文本口径里的 "careless" 对齐。
      enum: [concept_unclear, calculation_error, misreading, careless, knowledge_gap, other]
    ErrorIntent:
      type: string
      description: >-
        用户主观意图（D48）。review=想复习 / good=好题 / typical=典型 / doubtful=存疑。
        错题 = 有错因标记；star 题 = 只有意图标记、无错因；两者都进艾宾浩斯复习队列，
        但 mastery 与归因只消费「有错因」的。
      enum: [review, good, typical, doubtful]
    ErrorRecordCreate:
      type: object
      required: [subject, rawText]
      properties:
        subject:
          $ref: '#/components/schemas/Subject'
        rawText:
          type: string
          maxLength: 4000
          description: 错题原文（仅本地，永不出域）
        studentAnswer:
          type: string
          maxLength: 4000
          description: 学生作答
        correctAnswer:
          type: string
          maxLength: 4000
          description: 正确答案
        errorType:
          type: string
          description: 错因分类，如 concept/calculation/careless
        errorNote:
          type: string
          maxLength: 4000
          description: 自述错因
        pointIds:
          type: array
          description: 用户确认关联的知识点 ID
          items:
            type: string
        errorCause:
          $ref: '#/components/schemas/ErrorCause'
        intent:
          $ref: '#/components/schemas/ErrorIntent'
        sourceExamId:
          type: string
          description: 来源考试 ID（D49；错题可标记来自某次考试）
    ErrorRecordUpdate:
      type: object
      description: 错题局部更新，至少传一项
      minProperties: 1
      properties:
        errorType:
          type: string
        errorNote:
          type: string
          maxLength: 4000
        status:
          type: string
          enum: [open, resolved]
        pointIds:
          type: array
          items:
            type: string
        errorCause:
          $ref: '#/components/schemas/ErrorCause'
        intent:
          $ref: '#/components/schemas/ErrorIntent'
        sourceExamId:
          type: string
          nullable: true
          description: 来源考试 ID；传 null 表示清除关联
    LinkedPoint:
      type: object
      required: [pointId, confidence]
      properties:
        pointId:
          type: string
        name:
          type: string
        confidence:
          type: number
    ErrorRecord:
      type: object
      required: [errorId, subject, rawText, status, createdAt]
      properties:
        errorId:
          type: string
        subject:
          $ref: '#/components/schemas/Subject'
        rawText:
          type: string
          maxLength: 4000
        studentAnswer:
          type: string
          nullable: true
        correctAnswer:
          type: string
          nullable: true
        errorType:
          type: string
          nullable: true
        errorNote:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, resolved]
        points:
          type: array
          items:
            $ref: '#/components/schemas/LinkedPoint'
        createdAt:
          type: string
          format: date-time
        lastReviewedAt:
          type: string
          format: date-time
          nullable: true
        errorCause:
          allOf:
            - $ref: '#/components/schemas/ErrorCause'
          nullable: true
          description: 结构化错因（D48）；无错因即非「错题」（可能是只有意图的 star 题）
        intent:
          allOf:
            - $ref: '#/components/schemas/ErrorIntent'
          nullable: true
          description: 用户主观意图（D48）
        sourceExamId:
          type: string
          nullable: true
          description: 来源考试 ID（D49）
    ErrorBookList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ErrorRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorRecordDeleted:
      type: object
      required: [deleted, errorId]
      properties:
        deleted:
          type: boolean
        errorId:
          type: string
    ReviewSubmit:
      type: object
      required: [recallCorrect]
      properties:
        recallCorrect:
          type: boolean
          description: 这次复习是否回忆正确
    ReviewResult:
      type: object
      required: [correct, nextReviewAt, intervalDays]
      properties:
        correct:
          type: boolean
        nextReviewAt:
          type: string
          format: date
          description: 下一次复习日期（回忆正确按 1/2/4/7/15 递进；错误回到 1 天）
        intervalDays:
          type: integer
          description: 下次复习间隔天数
...